:sparkles: add POST method
The POST request method is now working. Tested in postman, it will
return a msg to confirm that the product has been added in json format.
The format of the post request MUST be in the next way:
    {
     "tipo": "kind of drink",
     "marca": "Brand",
     "tamano": "Size",
     "nombre": "Product name",
     "activo": "0" // Boolean 1 will show in the web, 0 wont.
     }

The fields "id"(autoincrement), "lastupdate" and "destacado"(boolean 0) will be added
automatically by the model.

Resolves #10