Some styles improves and documentation update
Some minor changes done to the css to show a full with code style.Added
some more information to the main document and finally a mini script in
javascript added to show the year in the main document.

This is the version 2.0 and will deploy the app for the first time.