Refactory of filters
The 3 diferents filters (like,status,category) have been refactored in just 1 method on the controller and now the controller just  call one model (getFilter)
to deal with filters. That saved a lot of redundant code and make it
easier to read. Some issues with the database must be solved as
explained in issue #21, however
this is a major update that I need to do with some more time