:sparkles: Secondary status querystring filter
This commit redo part of the getAll code, making it clean, spliting the
4 conditionals (depending of the number of keys on the $params array)
and adding more filter to check the query strings returned in $param.
If the query string contents 3 params if will return an error (only 2
params are allow), if the query string contents 2 params, it will check
that this params match with the $allowfilters (primary param) &
$allowedSecondFilter (secundary param). So far secondary param only
allow "status" but I might add "favorite" later on.The method then, will
check if the values of the query strings are not empty and finally will
check that "status" param only allow 0 or 1 as value. If everithing is
all right it will return the result.(The final result is not implemented
yet).