:recycle: Refactory of getAll method on products controller
The main method of the products controller has been refactored taking
out some repeat code to try to ajust to the DRY concept of coding. Also
a new array with the error msg has been created, and  the construct
method will created at the very beggining. Some functions on the error
library has been removed as now are in the main Class "products". I'll
review this later.