:bug: Fix an error made crash the Api when an item wasn't found in db
Thix hot fix, solves a problem when you put an item ID that wasn't in
the DB throwing an Slim error, now is showing the right response "400,
the item doesnt exist in the DB".

this commit also solve a wrong msg shown when the item ID wasn't
numeric, now show the right msg.

Last, it  clean some unnecesary code.